VOC-31232: Updated Bitrise and Fastfile. Renamed FeedbackSample to DigitalSample
