+++ ContactSurvey/src/main/AndroidManifest.xml
@@ -3,13 +3,13 @@
     package="com.verint.xm.contactSurvey" >
 
     <application
-        android:name="com.foresee.contactSurvey.CustomApplication"
+        android:name="com.verint.xm.contactSurvey.CustomApplication"
         android:label="@string/app_name"
         android:icon="@mipmap/ic_launcher"
         android:allowBackup="false">
 
         <activity
-            android:name="com.foresee.contactSurvey.MainActivity"
+            android:name="com.verint.xm.contactSurvey.MainActivity"
             android:label="@string/app_name"
             android:configChanges="orientation|screenSize|keyboard|keyboardHidden"
             android:theme="@style/AppTheme"
@@ -20,10 +20,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="com.foresee.contactSurvey.SetContactDetailsActivity"
+            android:name="com.verint.xm.contactSurvey.SetContactDetailsActivity"
             android:label="@string/set_contact_details"
             android:configChanges="orientation|screenSize|keyboard"
-            android:parentActivityName="com.foresee.contactSurvey.MainActivity"
+            android:parentActivityName="com.verint.xm.contactSurvey.MainActivity"
             android:theme="@style/AppTheme">
         </activity>
     </application>
