+++ FeedbackSample/src/androidTest/java/com/verint/xm/digitalSample/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.verint.xm.feedbackSample;
+package com.verint.xm.digitalSample;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
