+++ FeedbackSample/src/main/AndroidManifest.xml
@@ -1,15 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.verint.xm.feedbackSample" >
+    package="com.verint.xm.digitalSample" >
 
     <application
-        android:name="com.foresee.feedbackSample.FeedbackSampleApplication"
+        android:name="com.verint.xm.digitalSample.FeedbackSampleApplication"
         android:label="@string/app_name"
         android:icon="@mipmap/ic_launcher"
         android:allowBackup="false">
 
         <activity
-            android:name="com.foresee.feedbackSample.MainActivity"
+            android:name="com.verint.xm.digitalSample.MainActivity"
             android:label="@string/app_name"
             android:configChanges="orientation|screenSize|keyboard|keyboardHidden"
             android:theme="@style/AppTheme"
