+++ FeedbackSample/src/main/java/com/verint/xm/digitalSample/FeedbackSampleApplication.java
@@ -1,4 +1,4 @@
-package com.verint.xm.feedbackSample;
+package com.verint.xm.digitalSample;
 
 import android.app.Application;
