+++ FeedbackSample/src/main/java/com/verint/xm/digitalSample/MainActivity.java
@@ -1,4 +1,4 @@
-package com.verint.xm.feedbackSample;
+package com.verint.xm.digitalSample;
 
 import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
