+++ FeedbackSample/src/test/java/com/verint/xm/digitalSample/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.verint.xm.feedbackSample;
+package com.verint.xm.digitalSample;
