+++ LocalizationSample/src/main/AndroidManifest.xml
@@ -3,13 +3,13 @@
     package="com.verint.xm.localizationSample" >
 
     <application
-        android:name="com.foresee.localizationSample.CustomApplication"
+        android:name="com.verint.xm.localizationSample.CustomApplication"
         android:label="@string/app_name"
         android:allowBackup="false"
         android:icon="@mipmap/ic_launcher">
 
         <activity
-            android:name="com.foresee.localizationSample.MainActivity"
+            android:name="com.verint.xm.localizationSample.MainActivity"
             android:label="@string/app_name"
             android:configChanges="orientation|screenSize|keyboard|keyboardHidden"
             android:theme="@style/AppTheme"
